microservice/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
requests==2.31.0
matplotlib==3.8.2
python-dotenv==1.0.0

